clean up old things that I've advanced since writing
